+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The versioning scheme will follow the [semantic versioning](https://semver.org/)
 - On *smalldrop_bioprint* add Bioprinter class that will control the whole system. It shall provide the main API. Add tests.
 - On *smalldrop_bioprint* SystemState class subscribes to system working state topic.
 - On *smalldrop_bioprint* add node to instantiate the Bioprinter class and launch file.
+- On *smalldrop_bioprint* add system state machine.
+- On *smalldrop_bioprint* add init and shutdown routines for robot arm.
 
 ## 0.3.0 (2020-03-30)